+++ libs/opengl/include/mrpt/opengl/CText3D.h
@@ -97,7 +97,7 @@ class CText3D : public CRenderizable
 	CText3D(
 		const std::string& str = std::string(""),
 		const std::string& fontName = std::string("sans"),
-		const double scale = 1.0,
+		const float scale = 1.0,
 		const mrpt::opengl::TOpenGLFontStyle text_style = mrpt::opengl::NICE,
 		const double text_spacing = 1.5, const double text_kerning = 0.1);
