+++ libs/opengl/src/CTextMessageCapable.cpp
@@ -59,7 +59,7 @@ void CTextMessageCapable::render_text_messages(const int w, const int h) const
 										  : int(m_2D_text.second.y * h));
 
 		// Font size and family:
-		double font_size = 10;
+		float font_size = 10.0f;
 		string font_name = "sans";
 		TOpenGLFontStyle font_style = mrpt::opengl::FILL;
 		double font_spacing = 1.5;
@@ -188,7 +188,7 @@ bool CTextMessageCapable::updateTextMessage(
 void CTextMessageCapable::addTextMessage(
 	const double x_frac, const double y_frac, const std::string& text,
 	const mrpt::img::TColorf& color, const std::string& font_name,
-	const double font_size, const mrpt::opengl::TOpenGLFontStyle font_style,
+	const float font_size, const mrpt::opengl::TOpenGLFontStyle font_style,
 	const size_t unique_index, const double font_spacing,
 	const double font_kerning, const bool has_shadow,
 	const mrpt::img::TColorf& shadow_color)
